+++ src/Linode/LinodeClient.cs
@@ -6,6 +6,7 @@ namespace Linode;
 internal sealed class LinodeClient : ILinodeClient
 {
     public IDomainsOperation Domains { get; }
+    public IMaintenancePoliciesOperation MaintenancePolicies { get; }
     public INetworkTransferPricesOperation NetworkTransferPrices { get; }
     public IRegionsOperation Regions { get; }
     public ITagsOperation Tags { get; }
@@ -15,6 +16,7 @@ internal sealed class LinodeClient : ILinodeClient
         var httpConnection = new HttpConnection(httpClient);
 
         Domains = new DomainsOperation(httpConnection);
+        MaintenancePolicies = new MaintenancePoliciesOperation(httpConnection);
         NetworkTransferPrices = new NetworkTransferPricesOperation(httpConnection);
         Regions = new RegionsOperation(httpConnection);
         Tags = new TagsOperation(httpConnection);